wip tests + refactoring
