+++ error-reporting-csharp-dotnet-tool-tests/error-reporting-csharp-dotnet-tool-tests.csproj
@@ -1,9 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
+    <OutputType>Library</OutputType>
     <TargetFramework>net5.0</TargetFramework>
     <RootNamespace>error_reporting_csharp_dotnet_tool_tests</RootNamespace>
+    <ApplicationIcon />
+    <StartupObject />
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="xunit" Version="2.4.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\error-reporting-csharp-dotnet-tool\error-reporting-csharp-dotnet-tool.csproj" />
+  </ItemGroup>
 
 </Project>
