+++ error-reporting-csharp-dotnet-tool/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "error-reporting-csharp-dotnet-tool": {
       "commandName": "Project",
-      "commandLineArgs": "\"C:\\repos\\error-crawler-csharp\\error-reporting-dotnet-tool-testproject\\error-reporting-dotnet-tool-testproject.csproj\""
+      "commandLineArgs": "-t ECC -p \"C:\\repos\\error-crawler-csharp\\error-reporting-dotnet-tool-testproject\\error-reporting-dotnet-tool-testproject.csproj\""
     }
   }
 }
