+++ error-reporting-csharp-dotnet-tool/error-reporting-csharp-dotnet-tool.csproj
@@ -17,6 +17,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+	  <PackageReference Include="CommandLineParser" Version="2.8.0" />
 	  <PackageReference Include="Microsoft.Build.Locator" Version="1.4.1" />
 	  <PackageReference Include="Microsoft.CodeAnalysis" Version="3.9.0" />
 	  <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.9.0" />
